airflow-k8s/dags/requirements.txt
apache-airflow>=3.1.7
dbt-glue==1.10.15
dbt-core==1.10.15